+++ changelog.md
@@ -2,9 +2,9 @@
 
 ## Unreleased
 
-### Added
+### Changed
 
--   `options.formsRendererOrigin` as required option to Forms Class Constructor
+-   `forms.generateFormUrl(formId, externalId)` to return a promise that will resolve with the same value as before
 
 ## 0.1.0 (2018-07-11)